+++ sdk-lib/src/androidTest/java/cash/z/ecc/android/sdk/integration/SmokeTest.kt
@@ -36,7 +36,9 @@ class SmokeTest {
         )
         assertTrue(
             "Invalid CacheDB params dir",
-            wallet.initializer.rustBackend.pathParamsDir.endsWith("cache/params")
+            wallet.initializer.rustBackend.saplingParamDir.endsWith(
+                "no_backup/co.electricoin.zcash"
+            )
         )
     }